requirements.txt
# Dépendances de base
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.3

# Machine Learning
scikit-learn>=1.3.0

# NLP
nltk>=3.8.1
transformers>=4.36.2

# Deep Learning
--find-links https://download.pytorch.org/whl/torch_stable.html
torch>=2.1.0

# Autres dépendances
tqdm>=4.65.0
scipy>=1.11.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1

# Interface utilisateur
streamlit>=1.29.0
streamlit-chat>=0.1.1 